requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
motor==3.6.0
pymongo==4.9.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# HTTP Client
httpx==0.27.2

# OpenAI
openai==1.52.0

# Environment
python-dotenv==1.0.1

# Utilities
python-multipart==0.0.12
aiofiles==24.1.0

# CORS
starlette==0.38.6

# Google Cloud Vision
google-cloud-vision==3.7.2

